requirements.txt
flask
python-dotenv
requests
fastapi
uvicorn
pyngrok
pinecone